dataengineering/challenge-3/dashboard/requirements.txt
pymysql==0.9.3
Flask-MySQL==1.4.1